+++ src/test/java/seedu/address/commons/util/XmlUtilTest.java
@@ -19,8 +19,8 @@ import seedu.address.model.AddressBook;
 import seedu.address.storage.XmlAdaptedPerson;
 import seedu.address.storage.XmlAdaptedTag;
 import seedu.address.storage.XmlSerializableAddressBook;
-import seedu.address.testutil.AddressBookBuilder;
-import seedu.address.testutil.PersonBuilder;
+import seedu.address.testutil.builders.AddressBookBuilder;
+import seedu.address.testutil.builders.PersonBuilder;
 import seedu.address.testutil.TestUtil;
 
 public class XmlUtilTest {
