+++ src/test/java/seedu/address/logic/commands/DeleteCommandTest.java
@@ -6,9 +6,9 @@ import static org.junit.Assert.assertTrue;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandFailure;
 import static seedu.address.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static seedu.address.logic.commands.CommandTestUtil.showPersonAtIndex;
-import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
-import static seedu.address.testutil.TypicalIndexes.INDEX_SECOND_PERSON;
-import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
+import static seedu.address.testutil.testdata.TypicalIndexes.INDEX_FIRST_PERSON;
+import static seedu.address.testutil.testdata.TypicalIndexes.INDEX_SECOND_PERSON;
+import static seedu.address.testutil.testdata.TypicalPersons.getTypicalAddressBook;
 
 import org.junit.Test;
