+++ src/test/java/seedu/address/model/ModelManagerTest.java
@@ -3,8 +3,8 @@ package seedu.address.model;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_PERSONS;
-import static seedu.address.testutil.TypicalPersons.ALICE;
-import static seedu.address.testutil.TypicalPersons.BENSON;
+import static seedu.address.testutil.testdata.TypicalPersons.ALICE;
+import static seedu.address.testutil.testdata.TypicalPersons.BENSON;
 
 import java.nio.file.Paths;
 import java.util.Arrays;
@@ -14,7 +14,7 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
 import seedu.address.model.person.NameContainsKeywordsPredicate;
-import seedu.address.testutil.AddressBookBuilder;
+import seedu.address.testutil.builders.AddressBookBuilder;
 
 public class ModelManagerTest {
     @Rule
