+++ src/test/java/seedu/address/model/VersionedAddressBookTest.java
@@ -4,9 +4,9 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static seedu.address.testutil.TypicalPersons.AMY;
-import static seedu.address.testutil.TypicalPersons.BOB;
-import static seedu.address.testutil.TypicalPersons.CARL;
+import static seedu.address.testutil.testdata.TypicalPersons.AMY;
+import static seedu.address.testutil.testdata.TypicalPersons.BOB;
+import static seedu.address.testutil.testdata.TypicalPersons.CARL;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -14,7 +14,7 @@ import java.util.List;
 
 import org.junit.Test;
 
-import seedu.address.testutil.AddressBookBuilder;
+import seedu.address.testutil.builders.AddressBookBuilder;
 
 public class VersionedAddressBookTest {
