+++ src/test/java/seedu/address/model/person/NameContainsKeywordsPredicateTest.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 import org.junit.Test;
 
-import seedu.address.testutil.PersonBuilder;
+import seedu.address.testutil.builders.PersonBuilder;
 
 public class NameContainsKeywordsPredicateTest {
