+++ src/test/java/seedu/address/storage/XmlSerializableAddressBookTest.java
@@ -12,7 +12,7 @@ import org.junit.rules.ExpectedException;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.commons.util.XmlUtil;
 import seedu.address.model.AddressBook;
-import seedu.address.testutil.TypicalPersons;
+import seedu.address.testutil.testdata.TypicalPersons;
 
 public class XmlSerializableAddressBookTest {
