--- src/test/java/seedu/address/storage/XmlSerializableMakerManagerAdmins.java
@@ -1,4 +1,0 @@
-package seedu.address.storage;
-
-public class XmlSerializableMakerManagerAdmins {
-}
