+++ src/test/java/seedu/address/storage/XmlSerializableMakerManagerAdminsTest.java
@@ -1,0 +1,4 @@
+package seedu.address.storage;
+
+public class XmlSerializableMakerManagerAdminsTest {
+}
