+++ src/test/java/seedu/address/testutil/builders/MachineBuilder.java
@@ -1,4 +1,4 @@
-package seedu.address.testutil;
+package seedu.address.testutil.builders;
 
 
 import java.util.ArrayList;
